refactor(Main): MainViewModel 구현
- view 와 view model 이 1:1 대응이 되도록 변경
